% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Cartesian.R
\name{cartesian_2D}
\alias{cartesian_2D}
\title{Cartesian Product of Two Vectors}
\usage{
cartesian_2D(array_x, array_y)
}
\arguments{
\item{array_x}{A vector, indicating the first set.}

\item{array_y}{A vector, indicating the second set.}
}
\value{
A matrix of \code{length(array_x)} * \code{length(array_y)} rows
and two columns. Each row indicates an ordered pair.
}
\description{
Function \code{cartesian_2D} takes two vectors as input and
returns the two dimensional cartesian product.
}
\section{Comment}{

\code{cartesian_2D} is used internally in other function(s) of \pkg{ROCit}.
Works if matrix/data frames are passed as arguments. However,
returns might not be valid if arguments are not one dimensional.
}

\examples{
x <- seq(3)
y <- c(10,20,30)
cartesian_2D(x,y)


}
