% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aucCI.R
\name{ciAUC}
\alias{ciAUC}
\title{Confidence Interval of AUC}
\usage{
ciAUC(object, ...)
}
\arguments{
\item{object}{An object of class \code{"rocit"}, returned by
\code{\link{rocit}}.}

\item{...}{Arguments to be passed to methods.
See \code{\link{ciAUC.rocit}}.}
}
\description{
See \code{\link{ciAUC.rocit}}.
}
