% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rocCI.R
\name{ciROC}
\alias{ciROC}
\title{Confidence Interval of ROC curve}
\usage{
ciROC(object, ...)
}
\arguments{
\item{object}{An object of class \code{"rocit"}, returned by
\code{\link{rocit}}. Supports \code{"empirical"} and \code{"binormal"}
ROC curve.}

\item{...}{Arguments to be passed to methods.
See \code{\link{ciROC.rocit}}.}
}
\description{
See \code{\link{ciROC.rocit}}.
}
