% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gainstable.R
\name{gainstable}
\alias{gainstable}
\title{Gains Table for Binary Classifier}
\usage{
gainstable(...)
}
\arguments{
\item{...}{Arguments to be passed to methods.
See \code{\link{gainstable.default}}, \code{\link{gainstable.rocit}}.}
}
\description{
See \code{\link{gainstable.default}},
\code{\link{gainstable.rocit}}.
}
