% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/logitInvLogit.R
\name{invlogit-deprecated}
\alias{invlogit-deprecated}
\title{Logistic Transformation}
\usage{
invlogit(x)
}
\arguments{
\item{x}{A number or numeric vector from \eqn{(-Inf Inf)}.}
}
\value{
Logistic transformed value of \code{x}, \eqn{1/(1+exp(-x))}
}
\description{
\code{invlogit} returns the logistic transformed value.
}
\seealso{
\code{\link{ROCit-deprecated}}
}
\keyword{internal}
