% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ksplot.R
\name{ksplot}
\alias{ksplot}
\title{KS Plot}
\usage{
ksplot(object, ...)
}
\arguments{
\item{object}{An object of class \code{"rocit"}, returned by
\code{\link{rocit}} function.}

\item{...}{Arguments to be passed to methods.
See \code{\link{ksplot.rocit}}.}
}
\description{
See \code{\link{ksplot.rocit}}.
}
