% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/logitInvLogit.R
\name{logit-deprecated}
\alias{logit-deprecated}
\title{Log Odds of Probability}
\usage{
logit(x)
}
\arguments{
\item{x}{A number or numeric vector from \eqn{[0 1]}.}
}
\value{
Log-odds of \code{x}, \eqn{log(x/(1-x))}.
}
\description{
\code{logit} transforms probability value(s) into log-odds.
}
\seealso{
\code{\link{ROCit-deprecated}}
}
\keyword{internal}
