% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/measureit.R
\name{measureit}
\alias{measureit}
\title{Performance Metrics of Binary Classifier}
\usage{
measureit(...)
}
\arguments{
\item{...}{Arguments to be passed to methods.
See \code{\link{measureit.default}}, \code{\link{measureit.rocit}}.}
}
\description{
See \code{\link{measureit.default}},
\code{\link{measureit.rocit}}
}
