% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ROCit-defunct.R, R/getTPFPTNFN.R
\name{ROCit-defunct}
\alias{ROCit-defunct}
\alias{gettptnfpfn}
\title{Removed functions in package \pkg{ROCit}.}
\usage{
gettptnfpfn(x, depth, npos = sum(x), nneg = length(x) - npos)
}
\description{
The function(s) listed below are removed from
\pkg{ROCit}. Help pages for deprecated functions are
  available at \code{help("-defunct")}.
}
\keyword{internal}
