% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SummaryRocit.R
\name{summary.rocit}
\alias{summary.rocit}
\title{Summary of rocit object}
\usage{
\method{summary}{rocit}(object, ... = NULL)
}
\arguments{
\item{object}{An object of class \code{rocit}, returned by rocit function.}

\item{...}{\code{NULL}. Used for S3 generic/method consistency.}
}
\description{
Prints the summary of rocit object.
}
\examples{
data("Diabetes")
roc_empirical <- rocit(score = Diabetes$chol, class = Diabetes$dtest,
                       negref = "-")
# ---------------------
summary(roc_empirical)


}
\seealso{
\code{\link{rocit}}, \code{\link{print.rocit}}
}
